Move `get_considered_visits_table` to `BatchedMctsUtility`
Preparation to add `gumbel_explore` to `BatchedMctsAos`
